+++ packages/compiler-cli/index.ts
@@ -9,6 +9,7 @@ import {NodeJSFileSystem, setFileSystem} from './src/ngtsc/file_system';
 
 export {VERSION} from './src/version';
 
+export * from './src/transformers/jit_transforms';
 export * from './src/transformers/api';
 export * from './src/transformers/entry_points';
