+++ packages/compiler-cli/private/tooling.ts
@@ -15,9 +15,7 @@
 
 import ts from 'typescript';
 
-import {PartialEvaluator} from '../src/ngtsc/partial_evaluator';
-import {TypeScriptReflectionHost} from '../src/ngtsc/reflection';
-import {getDownlevelDecoratorsTransform, getInputSignalsMetadataTransform} from '../src/transformers/jit_transforms/index';
+import {angularJitApplicationTransform} from '../src/transformers/jit_transforms/index';
 
 /**
  * Known values for global variables in `@angular/core` that Terser should set using
@@ -34,47 +32,12 @@ export const GLOBAL_DEFS_FOR_TERSER_WITH_AOT = {
 };
 
 /**
- * JIT transform for Angular applications. Used by the Angular CLI for unit tests and
- * explicit JIT applications.
+ * JIT transform used by the Angular CLI.
  *
- * The transforms include:
- *
- *  - A transform for downleveling Angular decorators and Angular-decorated class constructor
- *    parameters for dependency injection. This transform can be used by the CLI for JIT-mode
- *    compilation where constructor parameters and associated Angular decorators should be
- *    downleveled so that apps are not exposed to the ES2015 temporal dead zone limitation
- *    in TypeScript. See https://github.com/angular/angular-cli/pull/14473 for more details.
- *
- *  - A transform for adding `@Input` to signal inputs. Signal inputs cannot be recognized
- *    at runtime using reflection. That is because the class would need to be instantiated-
- *    but is not possible before creation. To fix this for JIT, a decorator is automatically
- *    added that will declare the input as a signal input while also capturing the necessary
- *    metadata
+ * NOTE: Signature is explicitly captured here to highlight the
+ * contract various Angular CLI versions are relying on.
  */
-export function angularJitApplicationTransform(
-    program: ts.Program, isCore = false): ts.TransformerFactory<ts.SourceFile> {
-  const typeChecker = program.getTypeChecker();
-  const reflectionHost = new TypeScriptReflectionHost(typeChecker);
-  const evaluator = new PartialEvaluator(reflectionHost, typeChecker, null);
-
-  const downlevelDecoratorTransform = getDownlevelDecoratorsTransform(
-      typeChecker, reflectionHost, [], isCore,
-      /* enableClosureCompiler */ false);
-
-  const inputSignalMetadataTransform = getInputSignalsMetadataTransform(reflectionHost, isCore);
-
-  return (ctx) => {
-    return (sourceFile) => {
-      sourceFile = inputSignalMetadataTransform(ctx)(sourceFile);
-      sourceFile = downlevelDecoratorTransform(ctx)(sourceFile);
-
-      return sourceFile;
+export const constructorParametersDownlevelTransform =
+    (program: ts.Program, isCore = false): ts.TransformerFactory<ts.SourceFile> => {
+      return angularJitApplicationTransform(program, isCore);
     };
-  };
-}
-
-/**
- * Re-export for backwards compatibility.
- * The Angular CLI relies on this name for the transform.
- */
-export const constructorParametersDownlevelTransform = angularJitApplicationTransform;
