+++ packages/core/test/acceptance/input_signals/signal_inputs_test_compiler.ts
@@ -8,7 +8,7 @@
 
 import {PartialEvaluator} from '@angular/compiler-cli/src/ngtsc/partial_evaluator';
 import {TypeScriptReflectionHost} from '@angular/compiler-cli/src/ngtsc/reflection';
-import {getInputSignalsMetadataTransform} from '@angular/compiler-cli/src/transformers/jit_transforms';
+import {getInitializerApiJitTransform} from '@angular/compiler-cli/src/transformers/jit_transforms';
 import fs from 'fs';
 import ts from 'typescript';
 
@@ -30,7 +30,7 @@ async function main() {
   const evaluator = new PartialEvaluator(host, program.getTypeChecker(), null);
   const outputFile = ts.transform(
       program.getSourceFile(inputTsExecPath)!,
-      [getInputSignalsMetadataTransform(host, /* isCore */ false)], program.getCompilerOptions());
+      [getInitializerApiJitTransform(host, /* isCore */ false)], program.getCompilerOptions());
 
   await fs.promises.writeFile(
       outputExecPath, ts.createPrinter().printFile(outputFile.transformed[0]));
